refactor: extract match operations into wvImportMatchController
Move 9 match-operation handlers (getMatchStats, acceptMatch,
rejectMatch, rejectRemaining, acceptAndRejectRest, acceptBatchMatches,
getMatchTree, selectMapImage, markManualFix) out of the
worldViewImportController monolith into a dedicated
wvImportMatchController file. The monolith re-exports them.

Pure refactor — no behavior change, byte-identical function bodies.
Continues the spine domain-split (ADR-0009).
